1.0.16

+++ src/main/java/kiinse/plugins/ggo/gungalewiki/gui/items/CustomItem.java
@@ -179,16 +179,55 @@ public class CustomItem implements GuiItem {
                 }
             }
             if (clickType == ClickType.RIGHT) {
-                fromGui.delete();
-                new GuiBuilder(player)
-                        .setPage(0)
-                        .setItem(item)
-                        .setGui(Gui.FROMITEM)
-                        .setLastGui(fromGui)
-                        .setPageManager(new PageManager().setStackItems(GuiUtils.getItemsFromThis(item)))
-                        .setStringItem(item)
-                        .setName(config.getString(Config.MENU_FROMITEM_NAME))
-                        .open(player);
+                var pages = new PageManager().setStackItems(GuiUtils.getItemsFromThis(item));
+                var list = pages.getPageItemStackList(0);
+                if (!list.isEmpty()) {
+                    fromGui.delete();
+                    if (list.size() == 1) {
+                        var resultItem = CustomStack.byItemStack(list.get(0)).getId();
+
+                        var oresData = gunGaleWiki.getOresData();
+                        if (oresData.hasOre(resultItem)) return;
+                        if (oresData.hasDrop(resultItem)) {
+                            var pgs = new PageManager().setOreItems(oresData.getOresByDrop(resultItem));
+                            if (pgs.hasPage(0)) {
+                                new GuiBuilder(player)
+                                        .setItem(resultItem)
+                                        .setPage(0)
+                                        .setGui(Gui.ORES)
+                                        .setLastGui(fromGui)
+                                        .setPageManager(pgs)
+                                        .setStringItem(resultItem)
+                                        .setName(config.getString(Config.MENU_ORES_NAME))
+                                        .open(player);
+                            }
+                            return;
+                        }
+
+                        var pgs = new PageManager().setRecipes(GuiUtils.getRecipes(resultItem));
+                        if (pgs.hasPage(0)) {
+                            var isFurnace = pgs.getPageRecipe(0) instanceof FurnaceRecipe;
+                            new GuiBuilder(player)
+                                    .setPage(0)
+                                    .setGui(isFurnace ? Gui.FURNACE : Gui.WORKBENCH)
+                                    .setLastGui(fromGui)
+                                    .setPageManager(pgs)
+                                    .setStringItem(resultItem)
+                                    .setName(config.getString(isFurnace ? Config.MENU_FURNACE_NAME : Config.MENU_WORKBENCH_NAME))
+                                    .open(player);
+                        }
+                    } else {
+                        new GuiBuilder(player)
+                                .setPage(0)
+                                .setItem(item)
+                                .setGui(Gui.FROMITEM)
+                                .setLastGui(fromGui)
+                                .setPageManager(new PageManager().setStackItems(GuiUtils.getItemsFromThis(item)))
+                                .setStringItem(item)
+                                .setName(config.getString(Config.MENU_FROMITEM_NAME))
+                                .open(player);
+                    }
+                }
             }
         };
     }
